backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS & Rate Limiting
slowapi==0.1.9

# Scheduler
APScheduler==3.10.4
pytz==2024.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# AI APIs
openai==1.10.0
anthropic==0.18.1
google-generativeai==0.3.2

# Telegram
python-telegram-bot==20.7

# Notion (optional)
notion-client==2.2.1

# CSV Export
pandas==2.1.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Security Testing
bandit==1.7.6
safety==3.0.1

# Utilities
python-dateutil==2.8.2
